Moved the floating cursor stuff to a separate file
