+++ lib/src/re_editor.dart
@@ -16,6 +16,7 @@ import 'package:flutter/services.dart';
 import 'package:re_highlight/re_highlight.dart';
 import 'package:isolate_manager/isolate_manager.dart';
 
+part '_code_floating_cursor.dart';
 part '_code_autocomplete.dart';
 part '_code_editable.dart';
 part '_code_extensions.dart';
